User Restrictions
Now admin(or user_id = 10) will have access to all the pages. Otherwise,
all other users, when trying to access http://localhost:3000/users page,
will be redirect to the homepage. Since it is easier to debug with full
access on users, I'll keep the restriction short(so just the users
page). When the site is done, will need to add in more restrictions by
putting in the restricted urls.